Moved hard/softReclip on Cigars from CigarUtils to the class it is used in (AlignmendInterval). Kept visible (package rather than private) for the sake of testing, thus methods marked with "[at]VisibleForTesting".
