+++ src/main/java/org/broadinstitute/hellbender/utils/read/CigarUtils.java
@@ -370,88 +370,6 @@ public final class CigarUtils {
     }
 
     /**
-     * Returns a new cigar where any existing soft-clips are transformed into
-     * hard-clips.
-     * <p>
-     *     If the input cigar have a combination of hard and soft-clips then these are
-     *     combined into single hard-clip elements on either side (left and right).
-     * </p>
-     * <p>
-     *     The input cigar must be valid (see {@link Cigar#isValid}).
-     * </p>
-     *
-     * @param cigar the input cigar.
-     * @throws IllegalArgumentException if the input {@code cigar} is {@code null} or non-valid.
-     * @return never {@code null}.
-     */
-    public static Cigar hardReclip(final Cigar cigar) {
-        return softOrHardReclip(cigar, CigarOperator.H);
-    }
-
-    /**
-     * Returns a new cigar where any existing hard-clips are transformed into
-     * soft-clips.
-     * <p>
-     *     If the input cigar have a combination of hard and soft-clips then these are
-     *     combined into single soft-clip elements on either side (left and right).
-     * </p>
-     * <p>
-     *     The input cigar must be valid (see {@link Cigar#isValid}).
-     * </p>
-     *
-     * @param cigar the input cigar.
-     * @throws IllegalArgumentException if the input {@code cigar} is {@code null} or non-valid.
-     * @return never {@code null}.
-     */
-    public static Cigar softReclip(final Cigar cigar) {
-        return softOrHardReclip(cigar, CigarOperator.S);
-    }
-
-    /**
-     * Returns a cigar where all clips are transformed into either soft or hard clips.
-     *
-     * <p>
-     *     After such transformation, any resulting adjacent clips are folded into a single element.
-     * </p>
-     * @param cigar the input cigar.
-     * @param clipOperator the output cigar unified clipping operator (must be either {@link CigarOperator#S} or {@link CigarOperator#H}.
-     *
-     * @throws IllegalArgumentException if the input cigar cannot be {@code null}, or if it is invalid.
-     * @return the output cigar after the transformation.
-     */
-    private static Cigar softOrHardReclip(final Cigar cigar, final CigarOperator clipOperator) {
-        Utils.nonNull(cigar, "the input cigar cannot be null");
-        final List<CigarElement> elements = cigar.getCigarElements();
-        final int elementsSize = elements.size();
-        if (elementsSize < 1) { // the only operand cannot be a clip.
-            return new Cigar();
-        } else if (elements.get(0).getOperator().isClipping() || elements.get(elementsSize - 1).getOperator().isClipping()) {
-            final List<CigarElement> resultElements = new ArrayList<>(elements.size());
-            // Stack-like construction of the result-elements list
-            // where we fold-left the clips elements that we encounter
-            CigarElement lastElement = null; // caches the last cigar element added.
-            for (final CigarElement element : elements) {
-                final CigarOperator operator = element.getOperator();
-                if (operator.isClipping()) {
-                    if (lastElement != null && lastElement.getOperator().isClipping()) { // fold-left merging clipping operations.
-                        final int newLength = element.getLength() + lastElement.getLength();
-                        resultElements.set(resultElements.size() - 1, lastElement = new CigarElement(newLength, clipOperator));
-                    } else if (operator != clipOperator) { // cannot reuse the element instance as it has the "wrong" operator.
-                        resultElements.add(lastElement = new CigarElement(element.getLength(), clipOperator));
-                    } else { // has the right clip-operator and is the first in the run of clipping operations if any.
-                        resultElements.add(lastElement = element);
-                    }
-                } else { // is not a clipping, just add it as it is.
-                    resultElements.add(lastElement = element);
-                }
-            }
-            return new Cigar(resultElements);
-        } else {
-            return new Cigar(elements);
-        }
-    }
-
-    /**
      * Returns the length of the original read considering all clippings based on this cigar.
      * <p>
      *     The result of applying this method on a empty cigar is zero.
